backend/requirements.txt
fastapi>=0.115.2
uvicorn>=0.27.0
sqlmodel>=0.0.16
python-multipart>=0.0.18
python-dotenv>=1.0.0
requests>=2.32.2
aiofiles>=23.2.1
pydantic>=2.6.1
httpx>=0.27.0
PyJWT>=2.8.0
bcrypt>=4.0.0
sse-starlette>=2.0.0
huggingface_hub>=0.20.0
soundfile>=0.12.0

# mmgp for efficient memory management (alternative to BitsAndBytes)
mmgp>=3.7.0
accelerate>=1.12.0
vector_quantize_pytorch>=1.27.0
tokenizers>=0.22.0
transformers>=4.57.0

# heartlib — HeartMuLa core engine (music generation + lyrics transcription)
heartlib @ git+https://github.com/HeartMuLa/heartlib.git

# Vertex AI (Gemini + Veo 3.1 video generation)
google-cloud-aiplatform>=1.60.0
google-auth>=2.29.0
aiohttp>=3.9.0
